+++ src/module/applications/createActorDialog.ts
@@ -1,3 +1,4 @@
+import { ActorDataConstructorData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/actorData'
 import { IronswornActor } from '../actor/actor'
 import { IronswornSettings } from '../helpers/settings'
 
@@ -27,6 +28,7 @@ export class CreateActorDialog extends FormApplication<CreateActorDialogOptions>
 
     html.find('.ironsworn__character__create').on('click', (ev) => this._characterCreate.call(this, ev))
     html.find('.ironsworn__shared__create').on('click', (ev) => this._sharedCreate.call(this, ev))
+    html.find('.ironsworn__site__create').on('click', (ev) => this._siteCreate.call(this, ev))
   }
 
   async _characterCreate(ev: JQuery.ClickEvent) {
@@ -36,18 +38,28 @@ export class CreateActorDialog extends FormApplication<CreateActorDialogOptions>
     const table: any = await this._ironlanderNameTable()
     const drawResult = await table?.draw({ displayChat: false })
 
-    this._createWithFolder(drawResult.results[0]?.data.text, 'character')
+    this._createWithFolder(drawResult.results[0]?.data.text || 'Character', 'character', ev.currentTarget.dataset.img || undefined)
   }
 
   async _sharedCreate(ev: JQuery.ClickEvent) {
     ev.preventDefault()
-    this._createWithFolder('Shared', 'shared')
+    this._createWithFolder('Shared', 'shared', ev.currentTarget.dataset.img || undefined)
   }
 
-  async _createWithFolder(name: string, type: string) {
-    const data: any = {
+  async _siteCreate(ev: JQuery.ClickEvent) {
+    ev.preventDefault()
+    this._createWithFolder('Site', 'site', ev.currentTarget.dataset.img || undefined)
+  }
+
+  async _createWithFolder(name: string, type: 'character' | 'site' | 'shared', img: string) {
+    const data: ActorDataConstructorData & Record<string, any> = {
       name,
+      img,
       type,
+      token: {
+        displayName: CONST.TOKEN_DISPLAY_MODES.ALWAYS,
+        actorLink: true,
+      },
       folder: this.options.folder || undefined,
     }
     await IronswornActor.create(data, { renderSheet: true })