+++ src/module/item/delve-theme-domain/delvethemeitem.ts
@@ -1,0 +1,12 @@
+import { IronswornCharacterSheet } from "../../actor/sheets/charactersheet";
+import { BaseItem } from "../baseitem";
+
+export class DelveThemeItem extends BaseItem {
+  static entityName = 'delve-theme'
+
+  static activateActorSheetListeners(html: JQuery, sheet: IronswornCharacterSheet) {
+    super.activateActorSheetListeners(html, sheet)
+
+    // TODO?
+  }
+}