+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## In progress
 
+- Change compendium scene scale to 50
+- Delve sites ([#88](https://github.com/ben/foundry-ironsworn/pull/88))
+
 ## 1.2.0
 
 - Momentum does not apply to progress rolls ([#86](https://github.com/ben/foundry-ironsworn/pull/86))